% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{accordion_item}
\alias{accordion_item}
\title{'Accordion' items}
\usage{
accordion_item(
  title,
  ...,
  footer = NULL,
  class = "",
  collapsed = TRUE,
  parentId = rand_string(prefix = "accordion-"),
  itemId = rand_string(prefix = "accordion-item-"),
  style_header = NULL,
  style_body = NULL,
  root_path = template_root()
)
}
\arguments{
\item{title}{character title to show in the header}

\item{...}{body content}

\item{footer}{footer element, hidden if \code{NULL}}

\item{class}{the class of the item}

\item{collapsed}{whether collapsed at the beginning}

\item{parentId}{parent \code{\link{accordion}} id}

\item{itemId}{the item id}

\item{style_header, style_body}{'CSS' style of item header and body}

\item{root_path}{see \code{template_root}}
}
\value{
\code{'shiny.tag.list'} 'HTML' tags
}
\description{
'Accordion' items
}
\seealso{
\code{\link{accordion}}
}
