% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-adminlte.R
\name{adminlte}
\alias{adminlte}
\alias{adminlte_ui}
\alias{adminlte_sidebar}
\title{Generates 'AdminLTE' theme-related 'HTML' tags}
\usage{
adminlte_ui(root_path = template_root())

adminlte_sidebar(
  root_path = template_root(),
  settings_file = "modules.yaml",
  shared_id = rand_string(26)
)
}
\arguments{
\item{root_path}{the root path of the website project; see
\code{\link{template_settings}}}

\item{settings_file}{the settings file containing the module information}

\item{shared_id}{a shared identification by session to synchronize the
inputs; assigned internally.}
}
\value{
'HTML' tags
}
\description{
These functions should be called in 'HTML' templates.
Please see vignettes for details.
}
