% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{bslib_dependency}
\alias{bslib_dependency}
\title{Get Bootstrap 5 dependencies via \pkg{bslib}}
\usage{
bslib_dependency(...)
}
\arguments{
\item{...}{additional arguments passed to \code{bslib::bs_theme}}
}
\value{
An \code{\link[shiny]{tagList}} containing Bootstrap 5 dependencies
}
\description{
Returns Bootstrap 5 HTML dependencies provided by
\pkg{bslib}. Intended to be called from HTML templates so that
\code{headContent()} renders Bootstrap 5 CSS and JavaScript.
}
