% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawer.R
\name{drawer}
\alias{drawer}
\alias{drawer_open}
\alias{drawer_close}
\alias{drawer_toggle}
\title{Open, close, or toggle the drawer panel}
\usage{
drawer_open(session = shiny::getDefaultReactiveDomain())

drawer_close(session = shiny::getDefaultReactiveDomain())

drawer_toggle(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{shiny session}
}
\value{
No value is returned (called for side effect).
}
\description{
Send messages to the client to open, close, or toggle the
off-canvas drawer panel on the right side of the dashboard.
}
\examples{
server <- function(input, output, session){
  # Open the drawer
  drawer_open()

  # Close the drawer
  drawer_close()

  # Toggle the drawer
  drawer_toggle()
}

}
