% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{get_construct_string}
\alias{get_construct_string}
\title{Get \code{R} expression used to generate the 'HTML' tags}
\usage{
get_construct_string(x)
}
\arguments{
\item{x}{'HTML' tags}
}
\value{
Quoted \code{R} expressions that can generate the 'HTML' tags
}
\description{
This function only works on the elements generated by this
package
}
\examples{

x <- info_box("Message")
get_construct_string(x)

}
\seealso{
\code{\link{format_text_r}}
}
