% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{guess_body_class}
\alias{guess_body_class}
\title{Guess the 'AdminLTE' body class for modules, used internally}
\usage{
guess_body_class(cls)
}
\arguments{
\item{cls}{the class string of the \code{<body>} tag in \code{'index.html'}}
}
\value{
The proposed class for \code{<body>} tag
}
\description{
Guess the 'AdminLTE' body class for modules, used internally
}
