% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{html_asis}
\alias{html_asis}
\title{Escape HTML strings}
\usage{
html_asis(s, space = TRUE)
}
\arguments{
\item{s}{characters}

\item{space}{whether to also escape white space, default is true.}
}
\value{
An R string
}
\description{
Escape HTML strings so that they will be displayed
'as-is' in websites.
}
\examples{

html_asis("<a><----> <b>")

}
