% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{include_view}
\alias{include_view}
\title{Template function to include 'snippets' in the view folder}
\usage{
include_view(file, ..., .env = parent.frame(), .root_path = template_root())
}
\arguments{
\item{file}{files in the template \code{views} folder}

\item{...}{ignored}

\item{.env, .root_path}{internally used}
}
\value{
rendered 'HTML' segments
}
\description{
Store the reusing 'HTML' segments in the
\code{views} folder. This function should be used in the
\code{'index.html'} template
}
\examples{
\dontrun{
# in your 'index.html' file
<html>
<header>
{{ shidashi::include_view("header.html") }}
</header>

{{ shidashi::include_view("footer.html") }}
</html>
}
}
