% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{open_url}
\alias{open_url}
\title{Open a URL in a new browser tab}
\usage{
open_url(url, target = "_blank", session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{url}{character string, the URL to open}

\item{target}{the \code{window.open} target; default is
\code{"_blank"} (new tab)}

\item{session}{shiny session}
}
\value{
No value is returned (called for side effect).
}
\description{
Sends a message to the client to open the specified URL
in a new browser window/tab.
}
