% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-session.R
\name{register_global_reactiveValues}
\alias{register_global_reactiveValues}
\title{Register global reactive list}
\usage{
register_global_reactiveValues(
  name,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{name}{character, the key of the list}

\item{session}{shiny session}
}
\value{
A shiny \code{\link[shiny]{reactiveValues}} object
}
\description{
Creates or get reactive value list that is shared within the same
shiny session
}
