% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp-wrapper.R
\name{renderPlot2}
\alias{renderPlot2}
\title{Shiny render plot function with automated theme switcher}
\usage{
renderPlot2(expr, ..., env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr, env, quoted, ...}{passed to \code{\link[shiny]{renderPlot}}}
}
\value{
See \code{\link[shiny]{renderPlot}}
}
\description{
A wrapper around \code{\link[shiny]{renderPlot}}, but with themes
automatically set via \code{\link[graphics]{par}}; only supports
base plots. For \pkg{ggplot2}, please manually call
\code{\link{get_theme}} to get the theme.
}
\examples{

server <- function(input, output, session) {

  output$plot <- renderPlot2({
    plot(rnorm(100))
  })

}

}
