% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-viewer.R
\name{server_standalone_viewer}
\alias{server_standalone_viewer}
\title{Server function for the standalone viewer module}
\usage{
server_standalone_viewer(input, output, session, ...)
}
\arguments{
\item{input, output, session}{Shiny module server arguments}

\item{...}{ignored}
}
\description{
Internal server function used by the \code{standalone_viewer}
  hidden module. Retrieves the render function from the parent module session
  and assigns it to the viewer's output.
}
\keyword{internal}
