% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{show_ui_code}
\alias{show_ui_code}
\title{Used by demo project to show the generating code}
\usage{
show_ui_code(
  x,
  class = NULL,
  code_only = FALSE,
  as_card = FALSE,
  card_title = "",
  class_body = "bg-gray-70",
  width.cutoff = 80L,
  indent = 2,
  wrap = TRUE,
  args.newline = TRUE,
  blank = FALSE,
  copy_on_click = TRUE,
  ...
)
}
\arguments{
\item{x}{'HTML' tags generated by this package}

\item{class}{additional 'HTML' class}

\item{code_only}{whether to show code only}

\item{as_card}{whether to wrap results in \code{\link{card}}}

\item{card_title, class_body}{used by \code{\link{card}} if \code{as_card=TRUE}}

\item{width.cutoff, indent, wrap, args.newline, blank, copy_on_click, ...}{passed
to \code{\link{html_highlight_code}}}
}
\value{
'HTML' tags
}
\description{
Please write your own version. This function is designed for
demo-use only.
}
\seealso{
\code{html_highlight_code}
}
