% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{template_settings}
\alias{template_settings}
\alias{template_settings_set}
\alias{template_settings_get}
\alias{template_root}
\title{Configure template options that are shared across the sessions}
\usage{
template_settings

template_settings_set(...)

template_settings_get(name, default = NULL)

template_root()
}
\arguments{
\item{...}{key-value pair to set options}

\item{name}{character, key of the value}

\item{default}{default value if the key is missing}
}
\value{
\code{template_settings_get} returns the values represented by the
corresponding keys, or the default value if key is missing.
}
\description{
Configure template options that are shared across the sessions
}
\details{
The settings is designed to store static key-value pairs that
are shared across the sessions. The most important key is
\code{"root_path"}, which should be a path pointing to the template
folder.
}
\examples{

# Get current website root path

template_root()

}
