% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{use_template}
\alias{use_template}
\title{Download 'shidashi' templates from 'Github'}
\usage{
use_template(path, user = "dipterix", theme = "AdminLTE3", ...)
}
\arguments{
\item{path}{the path to create 'shidashi' project}

\item{theme}{the theme to download}

\item{name}{'Github' user name}
}
\value{
the target project path
}
\description{
Download 'shidashi' templates from 'Github'
}
\details{
To publish a 'shidashi' template, create a 'Github' repository
called \code{'shidashi-templates'}, or fork the \href{https://github.com/dipterix/shidashi-templates}{built-in templates}. The \code{theme} is the sub-folder
of the template repository.

An easy way to use a template in your project is through the 'RStudio'
project widget. In the 'RStudio' navigation bar, go to "File" menu,
click on the "New Project..." button, select the "Create a new project"
option, and find the item that creates 'shidashi' templates. Use the
widget to set up template directory.
}
