% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-item.R
\name{as_icon}
\alias{as_icon}
\title{Convert characters, shiny icons into 'fontawesome' 4}
\usage{
as_icon(icon = NULL, class = "fas")
}
\arguments{
\item{icon}{character or \code{\link[shiny]{icon}}}

\item{class}{icon class; change this when you are using 'fontawesome'
professional version. The choices are \code{'fa'} (compatible),
\code{'fas'} (strong), \code{'far'} (regular), \code{'fal'} (light),
and \code{'fad'} ('duotone').}
}
\value{
'HTML' tag
}
\description{
Convert characters, shiny icons into 'fontawesome' 4
}
\examples{

as_icon("bookmark", class = "far")
as_icon("bookmark", class = "fas")

# no icon
as_icon(NULL)

}
