% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{back_top_button}
\alias{back_top_button}
\title{'HTML' code to generate small back-to-top button}
\usage{
back_top_button(icon = "chevron-up", title = "Jump to")
}
\arguments{
\item{icon}{the icon for back-to-top button}

\item{title}{the expanded menu title}
}
\value{
'HTML' tags
}
\description{
This function is a template function that should be called
in 'HTML' templates before closing the \code{"</body>"} tag.
}
\examples{

back_top_button()
back_top_button("rocket")

}
