% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{init_app}
\alias{init_app}
\title{Initialize a shidashi application}
\usage{
init_app(env = parent.frame())
}
\arguments{
\item{env}{environment where the application is initialized into.}
}
\value{
Nothing.
}
\description{
Creates a fresh environment holding all per-application global stores
(session registry, per-module input registries, etc.).
This function is designed to be called from a \file{global.R} file
generated by \code{\link{render}()}, not for user's call.
}
\examples{

init_app(env = new.env())

}
