% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_info}
\alias{module_info}
\title{Obtain the module information}
\usage{
module_info(root_path = template_root(), settings_file = "modules.yaml")
}
\arguments{
\item{root_path}{the root path of the website project}

\item{settings_file}{the settings file containing the module information}
}
\description{
Obtain the module information
}
