% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-viz.R
\name{renderStreamViz}
\alias{renderStreamViz}
\title{Render a streaming widget}
\usage{
renderStreamViz(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An R expression that returns a \code{\link{stream_viz}} widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Logical. Whether \code{expr} is already quoted.}
}
\value{
A server-side render function for use with Shiny.
}
\description{
Server-side render function for \code{\link{streamVizOutput}}.
The expression should evaluate to a \code{\link{stream_viz}} object.
}
\seealso{
\code{\link{streamVizOutput}}, \code{\link{updateStreamViz}}
}
