% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-viz.R
\name{streamVizOutput}
\alias{streamVizOutput}
\title{Output placeholder for a streaming visualization widget}
\usage{
streamVizOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Character scalar. Output ID that matches the corresponding
\code{\link{renderStreamViz}} call.}

\item{width, height}{CSS width and height of the widget container.}
}
\value{
An HTML output element suitable for inclusion in a Shiny UI.
}
\description{
Use in a Shiny UI to reserve a slot for a multi-channel signal viewer
that is driven by binary stream data.
}
\seealso{
\code{\link{renderStreamViz}}, \code{\link{updateStreamViz}}
}
