% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{template_settings}
\alias{template_settings}
\alias{template_settings_set}
\alias{template_settings_get}
\alias{download_builtin_templates}
\alias{template_root}
\title{Configure template options that are shared across the sessions}
\format{
An object of class \code{list} of length 3.
}
\usage{
template_settings

template_settings_set(...)

template_settings_get(name, default = NULL)

download_builtin_templates()

template_root()
}
\arguments{
\item{...}{key-value pair to set options}

\item{name}{character, key of the value}

\item{default}{default value if the key is missing}
}
\description{
Configure template options that are shared across the sessions
}
\details{
The settings is designed to store static key-value pairs that
are shared across the sessions. The most important key is
\code{"root_path"}, which should be a path pointing to the template
folder.
}
\examples{

# Get current website root path

template_root()

}
\keyword{datasets}
